+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the "crates" extension will be documented in this file.
 - Comma handled for ranged versions [#117](https://github.com/serayuzgur/crates/issues/117)
 - Fix for comments with brackets inside arrays  [#115](https://github.com/serayuzgur/crates/pull/115)
 - ${version} is not being replaced for compatible decorator  [#118](https://github.com/serayuzgur/crates/issues/118)
+- Support alternative local index  [#109](https://github.com/serayuzgur/crates/issues/109)
 
 ## 0.5.5